+++ lib/static/static.dart
@@ -1,0 +1,11 @@
+final recipes = <Map<String,dynamic>>[
+  {
+    'id': 1,
+    'name': 'Pasta',
+    'description': 'Pasta with tomato sauce',
+  }, {
+    'id': 2,
+    'name': 'Pizza',
+    'description': 'Pizza with tomato sauce',
+  }
+];