+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to explain in detail changes made to the food structured app.
 Date: Nov XX, 2023
 * [NEW]
   * Blank calendar screen and drawer menu
+  * Recipes screen and edit recipe screen
 
 ## V 1.0.0
 Date: Nov 05, 2023